Create reusable date input, move primitives elements to separate directory
