+++ hat-menu/src/components/primitives/DateInput.tsx
@@ -1,20 +1,22 @@
 import type { ComponentPropsWithRef } from 'react';
-import { cn } from '../utils/styleUtils';
+import { cn } from '../../utils/styleUtils';
 
-type SelectProps = ComponentPropsWithRef<'select'> & {
+type DateInputProps = Omit<ComponentPropsWithRef<'input'>, 'type'> & {
     hasError?: boolean;
 };
 
-const Select = ({ className, hasError = false, ...props }: SelectProps) => {
+const DateInput = ({ className, hasError = false, ...props }: DateInputProps) => {
     const ariaInvalid = props['aria-invalid'] ?? (hasError || undefined);
 
     return (
-        <select
+        <input
+            type="date"
             aria-invalid={ariaInvalid}
             className={cn(
-                // NOTE: the styles should be kept quite matching the Input component
-                'w-full min-h-10 rounded-md border bg-white px-3 py-2 text-sm leading-5 text-gray-900',
+                'w-full max-w-38 min-h-10 rounded-md border bg-white px-3 py-2 text-sm leading-5 text-gray-900',
+                'placeholder:text-gray-400',
                 'disabled:cursor-not-allowed disabled:bg-gray-100 disabled:text-gray-500',
+                'read-only:bg-gray-50 read-only:text-gray-600',
                 hasError
                     ? 'border-red-500'
                     : 'border-gray-300',
@@ -25,4 +27,4 @@ const Select = ({ className, hasError = false, ...props }: SelectProps) => {
     );
 };
 
-export default Select;
+export default DateInput;
